+++ entrypoint.sh
@@ -4,7 +4,10 @@ set -e
 
 if [ "$(id -u)" -eq 0 ]; then
     nft add rule ip nat PREROUTING ip protocol tcp fib daddr type != local counter redirect to :9040 \
-        || echo 'failed to configure nftables for transparent proxy (missing CAP_NET_ADMIN?)'
+        || echo 'warning: failed to configure nftables for transparent proxy (missing CAP_NET_ADMIN?)'
+    nft add rule ip nat PREROUTING fib daddr type local udp dport 53 counter redirect to :9053 \
+        || echo 'warning: failed to configure nftables for DNS proxy' \
+                '(alternative for less flexible `docker run --publish 53:9053 ...`)'
     exec su -s /bin/sh tor -- "$0" "$@"
 fi